pyproject.toml
[tool.poetry]
name = "djadmin-detail-view"
version = "0.1.0"
description = ""
authors = ["Justin Louie <224840+nitsujri@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "djadmin_detail_view" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
django = ">=4.0"
django-environ = "^0.11.2"
whitenoise = "^6.6.0"
pytest-django = "^4.7.0"
django-webpack-loader = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest = "^7.4.3"
flake8 = "^6.1.0"
flake8-isort = "^6.1.0"
black = "^23.10.1"
factory-boy = "^3.3.0"
pytest-randomly = "^3.15.0"
python-dotenv = "^1.0.0"
django-pytest = "^0.2.0"
django-extensions = "^3.2.3"
uvicorn = "^0.27.1"
faker = "^23.2.1"
django-simple-history = "^3.4.0"
py-expect-to-change = { git = "ssh://git@github.com/jenfi-eng/py-expect-to-change.git", rev = "master" }
playwright = "^1.41.2"
django-sslserver = { git = "https://github.com/jenfi-eng/django-sslserver.git", rev = "master" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=example_project.settings_test --color=yes --code-highlight=yes"
python_files = ["tests.py", "test_*.py"]
norecursedirs = ["node_modules"]


# ==== Coverage ====
[tool.coverage.run]
include = ["nellie/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]


# ==== black ====
[tool.black]
line-length = 119
target-version = ['py311']


# ==== isort ====
[tool.isort]
profile = "black"
line_length = 119
known_first_party = ["nellie", "config"]
skip = ["venv/"]
skip_glob = ["**/migrations/*.py"]


# ==== mypy ====
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.environments.test"


# ==== PyLint ====
[tool.pylint.MASTER]
load-plugins = ["pylint_django"]
django-settings-module = "config.environments.local"

[tool.pylint.FORMAT]
max-line-length = 119

[tool.pylint."MESSAGES CONTROL"]
disable = ["missing-docstring", "invalid-name"]

[tool.pylint.DESIGN]
max-parents = 13

[tool.pylint.TYPECHECK]
generated-members = [
    "REQUEST",
    "acl_users",
    "aq_parent",
    "[a-zA-Z]+_set{1,2}",
    "save",
    "delete",
]

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002,T028,T003,H029,H016,H023,H021,H026,H013"
ignore_blocks = "raw"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2
